Type coercion is evil. And coffeescript doesn't help.
Because coffeescript forward declares all variables in a scope and t
hadn't been set, null was coerced to 0, causing t_ago to be a negative
number.  This resulted in no errors being thrown and redis not actually
removing anything from the sorted set.

But it's all just as well because ANNOUNCE_INTERVAL was 6 times smaller
than it should have been anyway!